backend/requirements.txt
alembic==1.17.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
greenlet==3.2.4
mako==1.3.10
markupsafe==3.0.3
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pypdfium2==5.0.0
python-dotenv==1.2.1
sqlalchemy==2.0.44
typing-extensions==4.15.0
google-generativeai==0.8.3
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
